% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abs}
\alias{abs}
\title{Absenteeism Study Data Set in a Sero Positive Population}
\format{
A data frame with 280 rows and 11 variables:
\describe{
\item{edad}{Age: 0 for no/low risk, 1 for risk factor present.}
\item{sexo}{Gender: 0 for no/low risk, 1 for risk factor present.}
\item{estado_civil}{Marital status: 0 for no/low risk, 1 for risk factor present.}
\item{ocupacion}{Occupation: 0 for no/low risk, 1 for risk factor present.}
\item{escolaridad}{Education level: 0 for no/low risk, 1 for risk factor present.}
\item{conteo_cd}{Count of chronic diseases: 0 for no/low risk, 1 for risk factor present.}
\item{carga_viral}{Viral load: 0 for no/low risk, 1 for risk factor present.}
\item{certificado_incapacidad}{Disability certificate: 0 for no/low risk, 1 for risk factor present.}
\item{dias_incapacidad}{Days of disability: 0 for no/low risk, 1 for risk factor present.}
\item{diagnostico_incapacidad}{Disability diagnosis: 0 for no/low risk, 1 for risk factor present.}
\item{causa_ausentismo}{Cause of absenteeism: 0 for no/low risk, 1 for risk factor present.}
}
}
\usage{
abs
}
\description{
Contains information about work absenteeism in a sero-positive population. It includes variables representing different risk factors for absenteeism, where 0 indicates no or low risk factors, and 1 indicates the presence of risk factors for each variable in the population.
}
\examples{
\donttest{
  # Load the dataset
  data("abs", package = "viruslearner")
  # Explore the dataset
  library(dplyr)
  dplyr::glimpse(abs)
}
}
\keyword{datasets}
