% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_fit.R
\name{cd_fit}
\alias{cd_fit}
\title{Fit and Evaluate Stacked Ensemble Model for CD4 Cell Count or Viral Load Outcome}
\usage{
cd_fit(obj_mod1, outcome, testdata, predicted, mode = TRUE)
}
\arguments{
\item{obj_mod1}{A stacked ensemble model.}

\item{outcome}{The name of the outcome variable.}

\item{testdata}{The test dataset for evaluating the ensemble.}

\item{predicted}{Column name of the predicted variable in a regression tidy format.}

\item{mode}{If TRUE, regression mode is used; if FALSE, classification mode is used (default is TRUE).}
}
\value{
A tibble containing the root mean squared error (RMSE) and coefficient of determination (R2) metrics.
}
\description{
This function fits a stacked ensemble model built with pre-processing and modeling workflows.
}
\examples{
\donttest{
if (!requireNamespace("baguette", quietly = TRUE)) {
install.packages("baguette")
}
library(baguette)
library(kernlab)
library(kknn)
library(ranger)
library(rules)
library(viruslearner)
data("obj_mod1", package = "viruslearner")
data("cd_test", package = "viruslearner")
testdata <- cd_test
outcome <- "cd_2023"
predicted <- ".pred"
obj_mod1 |> cd_fit(outcome, testdata, predicted, TRUE)
}
}
