% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cd_obj1}
\alias{cd_obj1}
\title{A stacked ensemble model of CD4 data}
\format{
A data frame with the following columns:
\describe{
\item{member}{Character vector, names of ensemble members}
\item{type}{Character vector, types of ensemble members}
\item{weight}{Numeric vector, weights of ensemble members}
}
}
\usage{
cd_obj1
}
\description{
This object represents a stacked ensemble model that retains 2 out of 9 possible candidate members.
It has a penalty of 0.0738619982207936 and a mixture of 1.
}
\details{
The 2 highest weighted members are:
\itemize{
\item \code{simple_Cubist_1_1}: Type - cubist_rules, Weight - 0.844
\item \code{simple_CART_bagged_1_1}: Type - bag_tree, Weight - 0.0907
}

Members have not yet been fitted with \code{fit_members()}.
}
\examples{
\donttest{
  # Load the object
  data("cd_obj1", package = "viruslearner")
  # Print the dataset
  print(cd_obj1)
}
}
\keyword{datasets}
