% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_stack.R
\name{cd_stack}
\alias{cd_stack}
\title{Plot Weighted Stacked Model}
\usage{
cd_stack(obj_mod1)
}
\arguments{
\item{obj_mod1}{A stacked ensemble model.}
}
\value{
A weighted stacked bar chart with text labels.
}
\description{
Generates a weighted stacked bar chart using the \code{autoplot} function from the \code{stacks} package and adds text labels using \code{geom_text} from \code{ggplot2}.
}
\examples{
\donttest{
if (!requireNamespace("baguette", quietly = TRUE)) {
install.packages("baguette")
}
library(baguette)
library(kernlab)
library(kknn)
library(ranger)
library(rules)
library(viruslearner)
data("obj_mod1", package = "viruslearner")
obj_mod1 |> cd_stack()
}
}
