% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cd_test}
\alias{cd_test}
\title{Viral Rates Dataset for Testing CD4 Counts Outcome}
\format{
A data frame with 9 rows and 17 variables:
\describe{
\item{cd_2019}{CD4 count in 2019.}
\item{vl_2019}{Viral load in 2019.}
\item{cd_2021}{CD4 count in 2021.}
\item{vl_2021}{Viral load in 2021.}
\item{cd_2022}{CD4 count in 2022.}
\item{vl_2022}{Viral load in 2022.}
\item{recovery_rate_2021}{CD4 count recovery rate from 2019 to 2021.}
\item{recovery_rate_2022}{CD4 count recovery rate from 2021 to 2022.}
\item{recovery_rate_2022_2yr}{CD4 count recovery rate from 2019 to 2022.}
\item{viral_rate_2021}{Viral load rate of change from 2019 to 2021.}
\item{viral_rate_2022}{Viral load rate of change from 2021 to 2022.}
\item{viral_rate_2022_2yr}{Viral load rate of change from 2019 to 2022.}
\item{adherence_1}{First principal component analysis scores representing adherence to ART.}
\item{adherence_2}{Second principal component analysis scores representing adherence to ART.}
\item{adherence_3}{Third principal component analysis scores representing adherence to ART.}
\item{adherence_4}{Fourth principal component analysis scores representing adherence to ART.}
\item{adherence_5}{Fifth principal component analysis scores representing adherence to ART.}
}
}
\source{
\url{https://cran.r-project.org/package=viraldomain}
}
\usage{
cd_test
}
\description{
The dataset contains testing observations about patients,
specifically their CD4 T cell counts (cd_2019, cd_2021, cd_2022) and viral
loads (vl_2019, vl_2021, vl_2022). For modeling patient recovery and viral
load persistence or suppression, column \code{cd_2022} is identified as the
outcome variable for patient recovery, and column \code{vl_2022} is identified as
the outcome variable for studying viral load persistence or suppression. The
dataset also contains information about variables related to adherence to
antiretroviral therapy(ART).
}
\examples{
\dontrun{
  # Load the dataset
  data("cd_test", package = "viruslearner")
  # Explore the dataset
  glimpse(cd_test)
}
}
\keyword{datasets}
