% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_nn.R
\name{class_nn}
\alias{class_nn}
\title{Train a neural network model and plot variable importance}
\usage{
class_nn(vip_featured, hiv_data, hu, plty, epo)
}
\arguments{
\item{vip_featured}{Outcome variable}

\item{hiv_data}{HIV data set}

\item{hu}{Hidden units in the model}

\item{plty}{Penalization}

\item{epo}{Training epochs in the model}
}
\value{
A variable importance plot
}
\description{
Train a neural network model and plot variable importance
}
\examples{
\donttest{
library(nnet)
data("mortality", package = "viruslearner")
vip_featured <- "V23"
hiv_data <- mortality
hu <- 16
plty <- 0.0000639
epo <- 720
set.seed(7)
class_nn(vip_featured, hiv_data, hu, plty, epo)
}
}
