% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obj_mod1}
\alias{obj_mod1}
\title{A stacked ensemble model of CD4 data}
\format{
A data frame with the following columns:
\describe{
\item{member}{Character vector, names of ensemble members}
\item{type}{Character vector, types of ensemble members}
\item{weight}{Numeric vector, weights of ensemble members}
}
}
\usage{
obj_mod1
}
\description{
This object represents a stacked ensemble model that retains 4 out of 9 possible candidate members.
It has a penalty of 0.0615848211066026 and a mixture of 1.
}
\details{
The 4 highest weighted members are:
\itemize{
\item \code{simple_Cubist_1_1}: Type - cubist_rules, Weight - 0.628
\item \code{simple_CART_bagged_1_1}: Type - bag_tree, Weight - 0.281
\item \code{full_quad_linear_reg_1_1}: Type - linear_reg, Weight - 0.0743
\item \code{normalized_SVM_poly_1_1}: Type - svm_poly, Weight - 0.0221
}

Members have not yet been fitted with \code{fit_members()}.
}
\examples{
\donttest{
  # Load the object
  data("obj_mod1", package = "viruslearner")
  # Print the dataset
  print(obj_mod1)
}
}
\keyword{datasets}
