% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obj_mod2}
\alias{obj_mod2}
\title{A screening model object of rcv data}
\format{
A data frame with the following columns:
\describe{
\item{wflow_id}{Character vector, unique identifier for each workflow}
\item{.config}{Character vector, configuration details for each workflow}
\item{.metric}{Character vector, performance metric (e.g., accuracy, ROC-AUC)}
\item{mean}{Numeric vector, mean performance metric}
\item{std_err}{Numeric vector, standard error of the mean}
\item{n}{Numeric vector, number of iterations or repetitions}
\item{preprocessor}{Character vector, preprocessor used in the workflow}
\item{model}{Character vector, type of model used in the workflow}
\item{rank}{Numeric vector, rank of the workflow based on performance}
}
}
\usage{
obj_mod2
}
\description{
This object represents a model screening generated using the viruslearner package. It contains information about the different competing models, their configurations, performance metrics, and rankings.
}
\details{
The model consists of multiple workflows (\code{wflow_id}) with different configurations (\code{.config}). For each workflow, there are metrics such as accuracy, ROC-AUC (\code{.metric}), mean performance (\code{mean}), standard error (\code{std_err}), number of iterations (\code{n}), preprocessor used (\code{preprocessor}), and the model type (\code{model}).

The model rankings (\code{rank}) provide insights into the relative performance of different workflows and configurations within the ensemble.
}
\examples{
\donttest{
  # Load the object
  data("obj_md2", package = "viruslearner")
  # Print the dataset
  print(obj_mod2)
}
}
\keyword{datasets}
