% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obj_mod5}
\alias{obj_mod5}
\title{A stacked ensemble model of prep data}
\format{
A data frame with the following columns:
\describe{
\item{member}{Character vector, names of ensemble members}
\item{type}{Character vector, types of ensemble members}
\item{weight}{Numeric vector, weights of ensemble members}
}
}
\usage{
obj_mod5
}
\description{
This object represents a stacked ensemble model that retains 2 out of 4 possible candidate members.
It has a penalty of 0.01 and a mixture of 1.
}
\details{
The 2 highest weighted members are:
\itemize{
\item \code{cate_neural_network_1_1}: Type - mlp, Weight - 5.65
\item \code{cate_svm_poly_1_1 }: Type - svm_poly, Weight - 2.40
}

Members have not yet been fitted with \code{fit_members()}.
}
\examples{
\donttest{
  # Load the object
  data("obj_mod5", package = "viruslearner")
  # Print the dataset
  print(obj_mod5)
}
}
\keyword{datasets}
