% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rcv}
\alias{rcv}
\title{Cardiovascular Risk Study Data Set in Sero Positive Population}
\format{
A data frame with 293 rows and 16 variables:
\describe{
\item{Sexo}{Sex of the individual (1 = male, 2 = female).}
\item{Edad}{Age of the individual.}
\item{Estado_civil}{Marital status (1 = single, 2 = married, 3 = divorced).}
\item{Escolaridad}{Education level (1 = primary, 2 = secondary, 3 = high school, 4 = technical, 5 = undergraduate, 6 = postgraduate).}
\item{HAS}{Hypertension status (1 = yes, 2 = no).}
\item{DM}{Diabetes mellitus status (1 = yes, 2 = no).}
\item{Tabaquismo}{Smoking status (1 = yes, 2 = no).}
\item{Cigarros}{Number of cigarettes smoked per day.}
\item{PAS}{Systolic blood pressure.}
\item{PAD}{Diastolic blood pressure.}
\item{HDL}{High-density lipoprotein cholesterol.}
\item{ColesterolT}{Total cholesterol level.}
\item{Trigliceridos}{Triglyceride level.}
\item{CargaViral}{Viral load status (1 = detectable, 2 = undetectable).}
\item{CD4}{CD4 count status (1 = low, 2 = normal).}
\item{RCV}{Cardiovascular risk status (1 = low, 2 = moderate/high).}
}
}
\usage{
rcv
}
\description{
Contains information from a cardiovascular risk study conducted in a seropositive population. It includes demographic variables such as sex (Sexo), age (Edad), marital status (Estado_civil), education level (Escolaridad), and health-related variables such as hypertension (HAS), diabetes mellitus (DM), smoking status (Tabaquismo), blood pressure measurements (PAS, PAD), lipid profile (HDL, ColesterolT, Trigliceridos), viral load (CargaViral), CD4 count (CD4), and the cardiovascular risk (RCV).
}
\examples{
\dontrun{
  # Load the dataset
  data("rcv", package = "viruslearner")
  # Explore the dataset
  library(dplyr)
  dplyr::glimpse(rcv)
}
}
\keyword{datasets}
