% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viruslearner-package.R
\docType{package}
\name{viruslearner-package}
\alias{viruslearner}
\alias{viruslearner-package}
\title{viruslearner: Ensemble Learning for HIV-Related Metrics}
\description{
Advanced statistical modeling techniques for ensemble learning, specifically tailored to the analysis of lymphocyte counts and viral load data in the context of HIV research. Empowering researchers and practitioners, this tool provides a comprehensive solution for tasks such as analysis, prediction and risk calculation related to key viral metrics. The package incorporates cutting-edge ensemble learning principles, inspired by model stacking techniques and adhering to tidy data principles, offering a robust and reproducible framework for HIV research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/juanv66x/viruslearner}
  \item Report bugs at \url{https://github.com/juanv66x/viruslearner/issues}
}

}
\author{
\strong{Maintainer}: Juan Pablo Acuña González \email{22253567@uagro.mx} (\href{https://orcid.org/0009-0003-6029-6560}{ORCID}) [copyright holder]

}
\keyword{internal}
