% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_ens.R
\name{vl_ens}
\alias{vl_ens}
\title{Viral load ensemble learning through stacking of models.}
\usage{
vl_ens(outcome, traindata, viralvars, logbase, seed, repetitions, gridsize)
}
\arguments{
\item{outcome}{The name of the outcome variable.}

\item{traindata}{The training dataset.}

\item{viralvars}{Vector of variable names related to viral data.}

\item{logbase}{The base for logarithmic transformations.}

\item{seed}{Seed for reproducibility.}

\item{repetitions}{Number of repetitions for cross-validation.}

\item{gridsize}{Size of the grid for hyperparameter tuning.}
}
\value{
A stacked ensemble model.
}
\description{
Stacking ensemble approach to combine predictions from various models,
allowing for grid search of tuning hyperparameters.
}
\examples{
\dontrun{
outcome <- "vl_2023"
traindata <- vl_train
viralvars <- c("vl_2019", "vl_2021", "vl_2022")
logbase <- 10
seed <- 1502
repetitions <- 2
gridsize <- 1
vl_ens(outcome, traindata, viralvars, logbase, seed, repetitions, gridsize)
}
}
