% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_stack.R
\name{vl_stack}
\alias{vl_stack}
\title{Viral load plot of blending coefficients for the stacking ensemble}
\usage{
vl_stack(outcome, traindata, viralvars, logbase, seed, repetitions, gridsize)
}
\arguments{
\item{outcome}{The outcome variable name (viral load).}

\item{traindata}{The training data set.}

\item{viralvars}{Vector of viral load variables.}

\item{logbase}{Logarithm base for viral load transformations.}

\item{seed}{Seed for reproducibility.}

\item{repetitions}{Number of repetitions for cross-validation.}

\item{gridsize}{Size of the grid for hyperparameter tuning.}
}
\value{
A plot of a data stack with multiple model definitions and candidate members.
}
\description{
A plot of the contribution of each model obtained using ensemble learning
through stacking of models and grid search for tuning hyperparameters of
viral load outcomes.
}
\examples{
\dontrun{
outcome <- "vl_2023"
traindata <- vl_train
viralvars <- c("vl_2019", "vl_2021", "vl_2022")
logbase <- 10
seed <- 1502
repetitions <- 2
gridsize <- 1
vl_stack(outcome, traindata, viralvars, logbase, seed, repetitions, gridsize)
}
}
