% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_ens.R
\name{cd_ens}
\alias{cd_ens}
\title{Build Custom Model}
\usage{
cd_ens(get_models = TRUE)
}
\arguments{
\item{get_models}{If TRUE, returns a list of predefined models without grid search or pre-processing (default is FALSE).}
}
\value{
Depending on the parameters, either a list of predefined models or a data frame with model evaluation results.
}
\description{
This function builds and evaluates custom machine learning models based on user-defined parameters
}
\examples{
# Get predefined models
cd_ens(get_models = TRUE)
}
