% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_fit.R
\name{cd_fit}
\alias{cd_fit}
\title{Fit and Evaluate Stacked Ensemble Model for CD4 Cell Count or Viral Load Outcome}
\usage{
cd_fit(obj_mod1, show_members = TRUE)
}
\arguments{
\item{obj_mod1}{A stacked ensemble model.}

\item{show_members}{If TRUE, outputs the retained members of the stacked ensemble model.}
}
\value{
A tibble containing the root mean squared error (RMSE) and coefficient of determination (R2) metrics,
or a summary of the model members.
}
\description{
This function fits a stacked ensemble model built with pre-processing and modeling workflows.
}
\examples{
\donttest{
library(Cubist)
library(kernlab)
library(kknn)
library(ranger)
library(rules)
data("fens_obj2", package = "viruslearner")
fens_obj2 |> cd_fit(show_members = TRUE)
}
}
