% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_stack.R
\name{cd_stack}
\alias{cd_stack}
\title{Plot Weighted Stacked Model}
\usage{
cd_stack(obj_mod1)
}
\arguments{
\item{obj_mod1}{A stacked ensemble model.}
}
\value{
A weighted stacked bar chart with text labels.
}
\description{
Generates a weighted stacked bar chart using the \code{autoplot} function
from the \code{stacks} package and adds text labels using \code{geom_text}
from \code{ggplot2}.
}
\examples{
\donttest{
library(Cubist)
library(kernlab)
library(kknn)
library(ranger)
library(rules)
data("fens_obj2", package = "viruslearner")
fens_obj2 |> cd_stack()
}
}
